HW3-FlightBooking/booking-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
pyjwt>=2.8.0
tenacity>=8.2.0
